Pull actions out into wake classes
