+++ game/rolesIndiv.js
@@ -1,46 +1,11 @@
-import Role, { Modifiers, Roles, Teams } from './role.js';
-import State, { CENTER_SIZE } from './state.js';
-import { assertUnreachable, makeList } from './utils.js';
+import Role, { Roles } from './role.js';
+import { assertUnreachable } from './utils.js';
+import {
+	ApprenticeSeerWake, DrunkWake, InsomniacWake, MasonWake, MinionWake, MysticWolfWake,
+	RevealerWake, RobberWake, SeerWake, SentinelWake, TroublemakerWake, WerewolfWake, WitchWake,
+} from './wakesIndiv.js';
 
 // #region Werewolves
-
-/**
- * Fetch a list of all the werewolf players at a current point in the game
- * @param {State} state
- * @returns {string[]}
- */
-function getAllWerewolves(state) {
-	const wolves = [];
-	for (let i = 0; i < state.numPlayers; i++) {
-		if (state.getPlayer(i).currentRole.killTeam === Teams.WEREWOLF) {
-			wolves.push(state.getName(i));
-		}
-	}
-	return wolves;
-}
-
-/**
- * Convert a list of werewolf names to a string describing all the werewolves.
- * @param {string[]} wolves
- * @returns {string}
- */
-function werewolfString(wolves) {
-	return `The werewolves are: ${makeList(wolves)}`;
-}
-
-/**
- * Returns the player names of the dream wolf, if one exists in the game
- * @param {State} state
- * @returns {string | null}
- */
-function getDreamWolf(state) {
-	for (let i = 0; i < state.numPlayers; i++) {
-		if (state.getPlayer(i).currentRole.role === Roles.DREAM_WOLF) {
-			return state.getName(i);
-		}
-	}
-	return null;
-}
 
 export class Werewolf extends Role {
 	constructor() {
@@ -48,32 +13,8 @@ export class Werewolf extends Role {
 			Roles.WEREWOLF,
 			'The bad guy',
 			'You will see who else started as a werewolf. Work together to fool the villagers.',
-			[2, 0],
+			[WerewolfWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		const wolves = getAllWerewolves(state);
-		giveInfo(werewolfString(wolves));
-		const dream = getDreamWolf(state);
-		if (dream !== null) {
-			giveInfo(`However, ${dream} is a dream wolf`);
-		}
-		if (wolves.length === 1) {
-			const pick = await pickCenters(1);
-			if (pick.length === 1) {
-				giveInfo(`The center card you picked was ${state.getCenter(pick[0]).roleName}`);
-			}
-		}
 	}
 }
 
@@ -83,36 +24,8 @@ export class MysticWolf extends Role {
 			Roles.MYSTIC_WOLF,
 			'The transcendent bad guy',
 			'You will see the other werewolves. You may also look at another player\'s card to help you fool the villagers.',
-			[2, 1],
+			[WerewolfWake, MysticWolfWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		const wolves = getAllWerewolves(state);
-		giveInfo(werewolfString(wolves));
-		const dream = getDreamWolf(state);
-		if (dream !== null) {
-			giveInfo(`However, ${dream} is a dream wolf and does not know the identity of the other werewolves.`);
-		}
-		if (wolves.length === 1) {
-			const pick = await pickCenters(1);
-			if (pick.length === 1) {
-				giveInfo(`The center card you picked was ${state.getCenter(pick[0]).roleName}`);
-			}
-		}
-		const view = await pickPlayers(1, false);
-		if (view.length === 1) {
-			giveInfo(`The card ${state.getName(view[0])} has is ${state.getPlayer(view[0]).currentRole.roleName}`);
-		}
 	}
 }
 
@@ -125,17 +38,6 @@ export class DreamWolf extends Role {
 			null,
 		);
 	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {}
 }
 
 export class Minion extends Role {
@@ -144,27 +46,8 @@ export class Minion extends Role {
 			Roles.MINION,
 			'You\'re a huge werewolf stan',
 			'You will see the other werewolves. You need to protect them, even if it costs you your life.',
-			[3],
+			[MinionWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		const ww = getAllWerewolves(state);
-		if (ww.length > 0) {
-			giveInfo(werewolfString(getAllWerewolves(state)));
-		} else {
-			giveInfo('There are no werewolves.');
-			giveInfo('If there are still no werewolves by the end of the night, then you must convince the villagers to kill a villager besides yourself to win.');
-		}
 	}
 }
 // #endregion
@@ -176,25 +59,8 @@ export class Sentinel extends Role {
 			Roles.SENTINEL,
 			'Protector',
 			'You may choose one other player to guard. That player\'s role will no longer be touched (may not be swapped, changed, and even the player themselves may not look at it)',
-			[0],
+			[SentinelWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		const protect = await pickPlayers(1, false);
-		if (protect.length === 1) {
-			const i = protect[0];
-			state.getPlayer(i).currentRole.modifiers.add(Modifiers.SENTINEL);
-		}
 	}
 }
 
@@ -204,31 +70,8 @@ export class Mason extends Role {
 			Roles.MASON,
 			'Twins',
 			'There are exactly two masons. You will get to see each other. If you do not see the other mason, the other mason started in the center.',
-			[4],
+			[MasonWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		const masons = [];
-		for (let i = 0; i < state.numPlayers; i++) {
-			if (state.getPlayer(i).currentRole.role === Roles.MASON) {
-				masons.push(state.getName(i));
-			}
-		}
-		if (masons.length === 1) {
-			giveInfo('You are the only mason. The other mason started in the center.');
-		} else {
-			giveInfo(`The masons are ${makeList(masons)}`);
-		}
 	}
 }
 
@@ -238,36 +81,8 @@ export class Seer extends Role {
 			Roles.SEER,
 			'Transcendent, or just a cheater',
 			'You may choose to either view two cards from the center or one other player\'s role.',
-			[5, 1],
+			[SeerWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		const choice = await pickChoice(['View two center cards', 'View one other player\'s card']);
-		if (choice === 0) {
-			// View two center
-			const cards = await pickCenters(2);
-			if (cards.length !== 0) {
-				const roles = cards.map((c) => state.getCenter(c).roleName);
-				giveInfo(`The cards you selected in the center were ${makeList(roles)}`);
-			}
-		} else if (choice === 1) {
-			// View one other player
-			const card = await pickPlayers(1, false);
-			if (card.length === 1) {
-				const pid = card[0];
-				giveInfo(`The card that ${state.getName(pid)} has is ${state.getPlayer(pid).currentRole.roleName}`);
-			}
-		}
 	}
 }
 
@@ -277,25 +92,8 @@ export class ApprenticeSeer extends Role {
 			Roles.APPRENTICE_SEER,
 			'Transcendent, but only kinda',
 			'You may view one card from the center.',
-			[5, 2],
+			[ApprenticeSeerWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		const cards = await pickCenters(1);
-		if (pickCenters.length === 1) {
-			const role = state.getCenter(cards[0]);
-			giveInfo(`The card you selected in the center was ${role.roleName}`);
-		}
 	}
 }
 
@@ -305,31 +103,8 @@ export class Robber extends Role {
 			Roles.ROBBER,
 			'Stealer of roles',
 			'You may steal (and view) one other player\'s role, taking on their role and giving them yours',
-			[6, 1],
+			[RobberWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		if (this.modifiers.has(Modifiers.SENTINEL)) {
-			giveInfo('Your role has been guarded by the sentinel. You will not be swapping roles tonight.');
-			return;
-		}
-
-		const cards = await pickPlayers(1, false);
-		if (cards.length === 1) {
-			const swapID = cards[0];
-			state.swap(id, swapID);
-			giveInfo(`Your new role is ${state.getPlayer(id).currentRole.roleName}`);
-		}
 	}
 }
 
@@ -339,37 +114,8 @@ export class Witch extends Role {
 			Roles.WITCH,
 			'Does some magic or something, idk',
 			'You may view one role from the center. If you do, you must swap it with any player of your choice.',
-			[6, 2],
+			[WitchWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		const cards = await pickCenters(1);
-		if (cards.length === 1) {
-			const centerCard = cards[0];
-			const fallbackPlayer = (() => {
-				const a = Math.floor(Math.random() * state.numPlayers);
-				if (state.getPlayer(a).currentRole.modifiers.has(Modifiers.SENTINEL)) {
-					// Fallback player cannot be sentinel
-					if (a === 0) { return 1; }
-					return a - 1;
-				}
-				return a;
-			})();
-			giveInfo(`The center card was ${state.getCenter(centerCard).roleName}. You must swap it with a player. If you do not select in time, it will be swapped with the following randomly selected player: ${state.getName(fallbackPlayer)}`);
-			const selectedPlayer = await pickPlayers(1, true);
-			const actualPlayer = selectedPlayer.length === 1 ? selectedPlayer[0] : fallbackPlayer;
-			state.swapCenter(actualPlayer, centerCard);
-		}
 	}
 }
 
@@ -379,24 +125,8 @@ export class Troublemaker extends Role {
 			Roles.TROUBLEMAKER,
 			'Screws with people',
 			'You may exchange the roles of two other players.',
-			[7],
+			[TroublemakerWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		const cards = await pickPlayers(2, false);
-		if (cards.length === 2) {
-			state.swap(cards[0], cards[1]);
-		}
 	}
 }
 
@@ -406,28 +136,8 @@ export class Drunk extends Role {
 			Roles.DRUNK,
 			'You have no idea what you\'re doing',
 			'You must pick a role from the center to exchange with your own. You do not get to see your new role.',
-			[8],
+			[DrunkWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		if (this.modifiers.has(Modifiers.SENTINEL)) {
-			giveInfo('Your role has been guarded by the sentinel. You will not be swapping roles tonight.');
-			return;
-		}
-
-		const cards = await pickCenters(1);
-		const card = cards.length === 1 ? cards[0] : Math.floor(Math.random() * CENTER_SIZE);
-		state.swapCenter(id, card);
 	}
 }
 
@@ -437,25 +147,8 @@ export class Insomniac extends Role {
 			Roles.INSOMNIAC,
 			'You can\'t fall asleep. Don\'t we all?',
 			'You get to wake up in the night and see your own role.',
-			[9],
+			[InsomniacWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		if (this.modifiers.has(Modifiers.SENTINEL)) {
-			giveInfo('Your role has been guarded by the sentinel. You will not see your current role tonight.');
-			return;
-		}
-		giveInfo(`Your current role is ${state.getPlayer(id).currentRole.roleName}`);
 	}
 }
 
@@ -465,33 +158,8 @@ export class Revealer extends Role {
 			Roles.REVEALER,
 			'Never learned how to keep your hands to yourself',
 			'You may reveal one other player\'s role. If it\'s on the villager team, it will also be revealed to all other players.',
-			[10],
+			[RevealerWake],
 		);
-	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {
-		const cards = await pickPlayers(1, false);
-		if (cards.length === 1) {
-			const pid = cards[0];
-			const pRole = state.getPlayer(pid).currentRole;
-
-			giveInfo(`The role ${state.getName(pid)} has is ${pRole.roleName}`);
-
-			if (pRole.winTeam === Teams.VILLAGER) {
-				pRole.modifiers.add(Modifiers.REVEALER);
-			} else {
-				giveInfo('Because this role is not on the villager team, it will not be shown to other players');
-			}
-		}
 	}
 }
 // #endregion
@@ -505,17 +173,6 @@ export class Tanner extends Role {
 			null,
 		);
 	}
-
-	/**
-	 * @param {(num: number, allowSelf: boolean) => Promise<number[]>} pickPlayers
-	 * Number of players and whether self selection is allowed to ids
-	 * @param {(num: number) => Promise<number[]>} pickCenters Number of cards to pick to ids
-	 * @param {(choices: string[]) => Promise<number>} pickChoice Array of choices to id of choice
-	 * @param {(msg: string) => void} giveInfo Show information to the player
-	 * @param {State} state Reference to the current game state
-	 * @param {number} id Current player ID
-	 */
-	async act(pickPlayers, pickCenters, pickChoice, giveInfo, state, id) {}
 }
 /**
  * Factory to construct a role object from a role ID
